Some changes necessary because of the difference between AD and standard classes and attributes
With the possible exception of a duplicate OID and a difference in the "name" attribute,
the changes are temporary. Attribute definitions will be returned to operational
when Samba provisioning is adjusted, and the difference to classes will be reverted
when the schema module is able to change the built-in classes at startup.

For now we need them so we can stop mapping classes on the Samba side.