Samba/AD specific schema handling
The purpose of this module is to create attribute and objects
from the classSchema and attributeSchema objects during provisioning
and at startup, so Samba does not have to map classes and attributes,
ald ultimately to allo updates of the schema by creating an entry in
cn=Schema partition.

It also creates index, refing and (wip) memberOf configurations based on
the schema entries flags.